refactor: rename toast components
