+++ apps/docs/src/.vitepress/components/IconLibraryItem.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 import OnyxIcon from "~components/OnyxIcon/OnyxIcon.vue";
-import { useToast } from "~components/OnyxToastProvider/useToast";
+import { useToast } from "~components/OnyxToast/useToast";
 import OnyxTooltip from "~components/OnyxTooltip/OnyxTooltip.vue";
 import type { EnrichedIcon } from "../utils-icons";
