+++ apps/docs/src/.vitepress/theme/TheLayout.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 import DefaultTheme from "vitepress/theme";
-import OnyxToastProvider from "~components/OnyxToastProvider/OnyxToastProvider.vue";
+import OnyxToast from "~components/OnyxToast/OnyxToast.vue";
 
 const { Layout } = DefaultTheme;
 </script>
@@ -8,7 +8,7 @@ const { Layout } = DefaultTheme;
 <template>
   <Layout>
     <template #layout-bottom>
-      <OnyxToastProvider />
+      <OnyxToast />
 
       <footer class="footer">
         <p class="footer__message">Released under the Apache-2.0 License.</p>
