+++ packages/sit-onyx/src/components/OnyxToast/OnyxToast.ct.tsx
@@ -1,79 +1,31 @@
-import { DENSITIES } from "../../composables/density";
-import { test } from "../../playwright/a11y";
-import { executeMatrixScreenshotTest } from "../../playwright/screenshots";
-import type { OnyxColor } from "../../types";
-import OnyxToast from "./OnyxToast.vue";
+import { expect, test } from "../../playwright/a11y";
+import { ONYX_BREAKPOINTS } from "../../types";
+import PlaywrightTestWrapper from "./PlaywrightTestWrapper.vue";
 
-const TOAST_COLORS = ["neutral", "danger", "warning", "success"] satisfies OnyxColor[];
+Object.entries(ONYX_BREAKPOINTS).forEach(([breakpoint, width]) => {
+  test(`should render (${breakpoint})`, async ({ mount, makeAxeBuilder, page }) => {
+    await page.setViewportSize({ width, height: 512 });
 
-test.beforeEach(async ({ page }) => {
-  await page.addStyleTag({
-    content: `.onyx-toast__progress-bar {
-      width: 100%
-    }`,
-  });
-});
+    // ARRANGE
+    const component = await mount(<PlaywrightTestWrapper />);
 
-test.describe("Screenshot tests", () => {
-  for (const mode of ["default", "clickable", "manual-close"] as const) {
-    executeMatrixScreenshotTest({
-      name: `Toast (${mode})`,
-      columns: DENSITIES,
-      rows: TOAST_COLORS,
+    // ASSERT
+    await expect(page).toHaveScreenshot(`breakpoint-${breakpoint}.png`);
+
+    // ACT
+    const accessibilityScanResults = await makeAxeBuilder()
       // TODO: remove when contrast issues are fixed in https://github.com/SchwarzIT/onyx/issues/410
-      disabledAccessibilityRules: ["color-contrast"],
-      component: (column, row) => (
-        <OnyxToast
-          headline="Test toast"
-          color={row}
-          density={column}
-          duration={mode === "manual-close" ? 0 : undefined}
-          clickable={mode === "clickable"}
-          // margin is used to not cut off the box shadow in the screenshot
-          style={{ margin: "1rem" }}
-        />
-      ),
-    });
-  }
-});
+      .disableRules(["color-contrast"])
+      .analyze();
 
-test.describe("Screenshot tests (description)", () => {
-  executeMatrixScreenshotTest({
-    name: "Toast (description)",
-    columns: DENSITIES,
-    rows: TOAST_COLORS,
-    // TODO: remove when contrast issues are fixed in https://github.com/SchwarzIT/onyx/issues/410
-    disabledAccessibilityRules: ["color-contrast"],
-    component: (column, row) => (
-      <OnyxToast
-        headline="Test toast"
-        color={row}
-        density={column}
-        description="Lorem ipsum dolor sit amet consectetur. Non in felis erat velit consectetur. Sed integer non hac viverra nibh vehicula risus ultrices. Molestie cras lobortis vitae gravida et ut. Turpis nisl pharetra amet ante eu sagittis sit elementum ut."
-        // margin is used to not cut off the box shadow in the screenshot
-        style={{ margin: "1rem" }}
-      />
-    ),
-  });
-});
+    // ASSERT
+    expect(accessibilityScanResults.violations).toEqual([]);
+    await expect(component).toContainText("Example toast 1");
 
-test.describe("Screenshot tests (truncation)", () => {
-  executeMatrixScreenshotTest({
-    name: "Toast (truncation)",
-    columns: ["default"],
-    rows: ["default"],
-    component: () => (
-      <OnyxToast
-        headline={"Test".repeat(32)}
-        description={"Test".repeat(96)}
-        // margin is used to not cut off the box shadow in the screenshot
-        style={{ margin: "1rem" }}
-      />
-    ),
-    beforeScreenshot: async (component) => {
-      await component.getByRole("paragraph").evaluate((element) => {
-        element.scrollBy({ top: element.scrollHeight });
-      });
-    },
+    // ACT
+    await component.getByLabel("Close").nth(4).click();
+
+    // ASSERT
+    await expect(component).not.toContainText("Example toast 1");
   });
 });
