+++ packages/sit-onyx/src/components/OnyxToast/OnyxToast.stories.ts
@@ -1,85 +1,56 @@
 import { defineStorybookActionsAndVModels } from "@sit-onyx/storybook-utils";
 import type { Meta, StoryObj } from "@storybook/vue3";
-import { defineIconSelectArgType } from "../../utils/storybook";
-import OnyxToast from "./OnyxToast.vue";
+import StorybookExample from "./StorybookExample.vue";
+import StorybookExampleSourceCode from "./StorybookExample.vue?raw";
 
 /**
- * Single toast component. Is mainly intended to be used with the [OnyxToastProvider](/docs/feedback-toastprovider--docs).
+ * The toast is used to display toast messages to the user.
+ * For a full list of toast examples, you can check out the [OnyxToastMessage](/docs/support-toastmessage--docs) component.
+ *
+ * ### Setup
+ *
+ * #### Step 1: Use the `<OnyxToast />` component
+ *
+ * Place the `<OnyxToast />` component once anywhere in your application (e.g. in your App.vue file).
+ * It will take care of displaying the toasts correctly.
+ *
+ * #### Step 2: Create the toast provider
+ *
+ * **If you are using the `createOnyx()` Vue plugin, you can skip this step since this will already be set up for you.**
+ *
+ * The toast works with Vue's [provide/inject](https://vuejs.org/guide/components/provide-inject) API so you need to create a toast provider once.
+ * To do so, add the following code to your `main.ts` file to set up the toasts globally:
+ *
+ * ```ts
+ * import { createToastProvider, TOAST_PROVIDER_INJECTION_KEY } from "sit-onyx";
+ *
+ * app.provide(TOAST_PROVIDER_INJECTION_KEY, createToastProvider());
+ * ```
+ *
+ * Afterwards, toasts can be shown using the `useToast()` composable as shown in the example below.
  */
-const meta: Meta<typeof OnyxToast> = {
-  title: "Support/Toast",
+const meta: Meta<typeof StorybookExample> = {
+  title: "Feedback/Toast",
   ...defineStorybookActionsAndVModels({
-    component: OnyxToast,
-    events: ["click", "close"],
-    argTypes: {
-      icon: defineIconSelectArgType(),
-    },
+    component: StorybookExample,
+    events: [],
+    decorators: [
+      (story) => ({
+        components: { story },
+        template: `<div style="height: 24rem;"> <story /> </div>`,
+      }),
+    ],
   }),
+  parameters: {
+    docs: {
+      source: {
+        code: StorybookExampleSourceCode.replace('from "../.."', 'from "sit-onyx"'),
+      },
+    },
+  },
 };
 
 export default meta;
-type Story = StoryObj<typeof OnyxToast>;
+type Story = StoryObj<typeof StorybookExample>;
 
-export const Default = {
-  args: {
-    headline: "Example toast",
-    description: "Lorem ipsum dolor sit amet consectetur. Non in felis erat velit consectetur.",
-  },
-} satisfies Story;
-
-export const Danger = {
-  args: {
-    ...Default.args,
-    color: "danger",
-  },
-} satisfies Story;
-
-export const Warning = {
-  args: {
-    ...Default.args,
-    color: "warning",
-  },
-} satisfies Story;
-
-export const Success = {
-  args: {
-    ...Default.args,
-    color: "success",
-  },
-} satisfies Story;
-
-export const Clickable = {
-  args: {
-    ...Default.args,
-    clickable: true,
-  },
-} satisfies Story;
-
-export const ManualClose = {
-  args: {
-    ...Default.args,
-    duration: 0,
-  },
-} satisfies Story;
-
-/**
- * This example shows a toast with a very long description that is limited to 3 lines.
- * The user can scroll to see the rest of the description.
- */
-export const LongDescription = {
-  args: {
-    ...Default.args,
-    description:
-      "Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.",
-  },
-} satisfies Story;
-
-/**
- * This example shows a toast with the minimal amount of data.
- */
-export const MinimalData = {
-  args: {
-    headline: Default.args.headline,
-    icon: false,
-  },
-} satisfies Story;
+export const Default = { args: {} } satisfies Story;
